Create backend updateUser by admin functionality #32

+++ react-frontend/src/screens/UserListScreen.js
@@ -32,7 +32,9 @@ function UserListScreen({ history }) {
     }, [dispatch, history, successDelete])
 
     const deleteHandler = (id) => {
-        dispatch(deleteUser(id))
+        if (window.confirm("Are you sure you want to delete this user?")) {
+            dispatch(deleteUser(id))
+        }
     }
 
     return (
